[#8] header에 sidebar 로직 추가 구현 완료
