+++ src/components/header/SidebarData.tsx
@@ -1,0 +1,41 @@
+import {
+  Home,
+  Newspaper,
+  Task,
+  Chat,
+  GroupAdd,
+  Assignment,
+} from "@mui/icons-material";
+
+export const SidebarData = [
+  {
+    title: "Home",
+    path: "/",
+    icon: <Home />,
+  },
+  {
+    title: "Issues",
+    path: "/job-news",
+    icon: <Newspaper />,
+  },
+  {
+    title: "채용공고",
+    path: "/job-posting",
+    icon: <Task />,
+  },
+  {
+    title: "Job담",
+    path: "/jobdam",
+    icon: <Chat />,
+  },
+  {
+    title: "GaJob-study",
+    path: "/gajob-study",
+    icon: <GroupAdd />,
+  },
+  {
+    title: "Portfolio",
+    path: "/portfolio",
+    icon: <Assignment />,
+  },
+];
