+++ src/components/slider/SlideItem.tsx
