+++ src/components/slider/index.tsx
