+++ src/pages/JobPosting.tsx
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+const JobPostingWrapper = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 50px;
+  height: 70vh;
+`;
+
+export const JobPosting = () => {
+  return <JobPostingWrapper>JobPosting</JobPostingWrapper>;
+};
