+++ src/pages/Portfolio.tsx
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+const PortfolioWrapper = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 50px;
+  height: 70vh;
+`;
+
+export const Portfolio = () => {
+  return <PortfolioWrapper>Portfolio</PortfolioWrapper>;
+};
